+++ app/controllers/admin/claims_controller.rb
@@ -13,4 +13,13 @@ class Admin::ClaimsController < Admin::BaseAdminController
   def show
     @claim = Claim.find(params[:id])
   end
+
+  def payroll
+    claims = Claim.includes(eligibility: [:claim_school, :current_school]).awaiting_approval.order(:submitted_at)
+    csv = PayrollDataCsv.new(claims)
+
+    respond_to do |format|
+      format.csv { send_file csv.file, type: "text/csv", filename: "payroll_data.csv" }
+    end
+  end
 end