+++ spec/features/admin_csv_download_spec.rb
@@ -20,6 +20,19 @@ RSpec.feature "Download CSV of claims" do
       csv = CSV.parse(body)
       expect(csv.count).to eq(submitted_claims.count + 1)
     end
+
+    scenario "User downloads a CSV for the payroll provider" do
+      submitted_claims = create_list(:claim, 5, :submittable, submitted_at: DateTime.now)
+      create_list(:claim, 2, :submittable)
+
+      expect(page).to have_link(nil, href: payroll_admin_claims_path(format: :csv))
+      click_on "Download payroll data"
+
+      expect(page.response_headers["Content-Type"]).to eq("text/csv")
+
+      csv = CSV.parse(body)
+      expect(csv.count).to eq(submitted_claims.count + 1)
+    end
   end
 
   context "User is logged in as a support user" do
@@ -33,6 +46,15 @@ RSpec.feature "Download CSV of claims" do
       expect(page).to_not have_link(nil, href: admin_claims_path(format: :csv))
 
       visit admin_claims_path(format: :csv)
+
+      expect(page.status_code).to eq(401)
+      expect(page).to have_content("Not authorised")
+    end
+
+    scenario "User cannot download a CSV for the payroll provider" do
+      expect(page).to_not have_link(nil, href: payroll_admin_claims_path(format: :csv))
+
+      visit payroll_admin_claims_path(format: :csv)
 
       expect(page.status_code).to eq(401)
       expect(page).to have_content("Not authorised")